apiMiddl/requirements.txt
Django==4.1
djangorestframework==3.13.1
django-oauth-toolkit==2.1.0
django-filter==22.1
requests==2.22.0
tableauserverclient==0.9
pandas==1.4.2
minio==5.0.7
termcolor==1.1.0
oauthenticator==15.0.1
docker==5.0.3
scikit-learn==1.1.2
coreapi==2.3.3